+++ src/main/java/com/yanny/ytech/generation/YTechItemModelsProvider.java
@@ -32,6 +32,7 @@ class YTechItemModelsProvider extends ItemModelProvider {
         clayBucketItem(YTechItems.WATER_CLAY_BUCKET);
 
         simpleItem(YTechItems.ANTLER);
+        simpleItem(YTechItems.BONE_NEEDLE);
         simpleItem(YTechItems.BREAD_DOUGH);
         simpleItem(YTechItems.BRICK_MOLD);
         simpleItem(YTechItems.CLAY_BUCKET);