+++ src/main/java/com/yanny/ytech/registration/YTechItems.java
@@ -34,6 +34,7 @@ public class YTechItems {
 
     public static final RegistryObject<Item> ANTLER = ITEMS.register("antler", YTechItems::simpleItem);
     public static final RegistryObject<Item> BASKET = ITEMS.register("basket", BasketItem::new);
+    public static final RegistryObject<Item> BONE_NEEDLE = ITEMS.register("bone_needle", () -> new ToolItem(Tiers.WOOD, false, new Item.Properties().durability(5).setNoRepair()));
     public static final RegistryObject<Item> BREAD_DOUGH = ITEMS.register("bread_dough", YTechItems::simpleItem);
     public static final RegistryObject<Item> BRICK_MOLD = ITEMS.register("brick_mold", () -> new Item(new Item.Properties().durability(256)));
     public static final RegistryObject<Item> CLAY_BUCKET = ITEMS.register("clay_bucket", () -> new ClayBucketItem(() -> Fluids.EMPTY, new Item.Properties().stacksTo(8)));